fix bug(input must unique & !specialchar)

+++ src/pages/anime-detail.js
@@ -6,6 +6,8 @@ import { useContext, useEffect } from "react";
 import { AnimeContext } from "./ContextAnime";
 import Button from "../component/button";
 import IncludeCollection from "../component/collection";
+import CollectionDetail from "./collection-detail";
+import Loading from "../component/loading";
 
 const AnimeDetail = () => {
   let { id } = useParams();
@@ -35,26 +37,60 @@ const AnimeDetail = () => {
   });
 
   const addNewCollection = async (data) => {
-    let colectionname = prompt("Collection Name");
-    let newData = {
-      name: colectionname,
-      data: [data],
-    };
-    localStorage.setItem("newCollection", JSON.stringify(newData));
+    let colectionname = prompt(
+      "Please input your Collection Name. Your Collection Name must unique and doesn't have special character"
+    );
+    //check collection name is already added
+    let checkName = list.some((element) => element.name === colectionname);
 
-    // Parse any JSON previously stored in listCollection
-    var existinglist = JSON.parse(localStorage.getItem("listCollection"));
-    if (existinglist == null) existinglist = [];
+    //checkSpecialChar
+    function containsSpecialChars(colectionname) {
+      const specialChars = /[`!@#$%^&*()_+\-=\[\]{};':"\\|,.<>\/?~]/;
+      return specialChars.test(colectionname);
+    }
+    containsSpecialChars(colectionname);
 
-    let newList = JSON.parse(localStorage.getItem("newCollection"));
+    //ifspecialchar
+    if (containsSpecialChars(colectionname) === true) {
+      colectionname = prompt(
+        "Your Collection Name have special character, Please add another collection name"
+      );
+    }
 
-    localStorage.setItem("newdata", JSON.stringify(newList));
-    // Save listCollection back to local storage
-    existinglist.push(JSON.parse(localStorage.getItem("newCollection")));
-    localStorage.setItem("listCollection", JSON.stringify(existinglist));
+    //if collcetionname is null, show prompt again
+    while (colectionname !== null && colectionname === "") {
+      colectionname = prompt("Collection Name");
+    }
+
+    //if collectionname is already added, call another prompt message
+    if (checkName === true) {
+      colectionname = prompt(
+        "Collection name is already added. Please add another collection name"
+      );
+    }
+
+    //If collection name  input isn't null and name is unique, add collection
+    if (colectionname !== null && checkName === false) {
+      let newData = {
+        name: colectionname,
+        data: [data],
+      };
+      localStorage.setItem("newCollection", JSON.stringify(newData));
+
+      // Parse any JSON previously stored in listCollection
+      var existinglist = JSON.parse(localStorage.getItem("listCollection"));
+      if (existinglist == null) existinglist = [];
+
+      let newList = JSON.parse(localStorage.getItem("newCollection"));
+
+      localStorage.setItem("newdata", JSON.stringify(newList));
+      // Save listCollection back to local storage
+      existinglist.push(JSON.parse(localStorage.getItem("newCollection")));
+      localStorage.setItem("listCollection", JSON.stringify(existinglist));
+    }
   };
 
-  if (loading) return "Loading...";
+  if (loading) return <Loading />;
   if (error) return `Error! ${error.message}`;
 
   return (
@@ -81,13 +117,11 @@ const AnimeDetail = () => {
         </div>
 
         <div className="test">
-          <div>This item is in collection :</div>
           {anime.length !== 0 ? (
-            (anime || []).map((item, index) => (
-              <>
-                <IncludeCollection item={item} index={index} />
-              </>
-            ))
+            <>
+              <div>This item is in collection :</div>
+              <IncludeCollection />
+            </>
           ) : (
             <div style={{ color: "red" }}>Not added to any collection yet</div>
           )}
